telemtry_F1/dashboard/requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
pandas==2.2.1
plotly==5.19.0